Move saved queries into the Keen namespace
- Move saved queries into the keen folder
- Update tests
- Move then use some common functions in client
